[[FIX]] Correct parsing of `let` token
Tolerate the use of `let` as an Identifier outside of strict mode.

This change makes error messages less helpful when the input code
contains the non-standard "`let` expression" language extension but
omits the required `moz` option. However, this change does not effect
JSHint's ability to parse that language extension and is therefore
compatible with prior releases of JSHint.